Factor out utility code from PipelineMainTest
